Simulation of HiC and WGS works

+++ src/simulation/wgsim.sh
@@ -15,12 +15,18 @@ template_count=$(grep -e "${contig}"$'\t' ${all_stat} | awk '{print int($3/4);}'
 echo $chrom $contig $template_count
 echo $index_start
 
-rm ${outdir}/${chrom}_1.fq
+if [ -f ${outdir}/${chrom}_1.fq ];
+then
+	rm ${outdir}/${chrom}_1.fq
+fi
+
 mkfifo ${outdir}/${chrom}_1.fq
 sed -n 1~4p ${outdir}/${chrom}_1.fq | python -m svphase.scripts.fq_to_predat ${index_start} | gzip 1>${outdir}/${chrom}.predat.gz &
 
 /home/anand/software/samtools-0.1.18/misc/wgsim -N ${template_count} -1 100 -2 100 ${fa} ${outdir}/${chrom}_1.fq /dev/null >${outdir}/${chrom}.out;
 sleep 5;
+wait
+rm ${outdir}/${chrom}_1.fq
 
 
 #python -m svphase.scripts.create_fa_from_sv ${prefix}.${i}.${letter}.bed;
